+++ app/mailers/admin_mailer.rb
@@ -50,6 +50,16 @@ class AdminMailer < ApplicationMailer
     mail(to: recipient.email, subject: @notification.title)
   end
 
+  def businesses_hiring_invoice_request
+    @notification = params[:record].to_notification
+    recipient = params[:recipient]
+
+    @form = @notification.form
+    @business = @form.business
+
+    mail(to: recipient.email, subject: @notification.title)
+  end
+
   def new_hired_form
     @notification = params[:record].to_notification
     recipient = params[:recipient]