+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_03_19_212906) do
+ActiveRecord::Schema[7.0].define(version: 2023_03_23_132220) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_stat_statements"
   enable_extension "plpgsql"
@@ -87,7 +87,22 @@ ActiveRecord::Schema[7.0].define(version: 2023_03_19_212906) do
     t.string "website"
     t.string "contact_role"
     t.boolean "invisible", default: false, null: false
+    t.boolean "survey_request_notifications", default: true
     t.index ["user_id"], name: "index_businesses_on_user_id"
+  end
+
+  create_table "businesses_hiring_invoice_requests", force: :cascade do |t|
+    t.bigint "business_id", null: false
+    t.text "billing_address", null: false
+    t.string "developer_name", null: false
+    t.date "start_date", null: false
+    t.integer "annual_salary", null: false
+    t.integer "employment_type", null: false
+    t.string "position"
+    t.text "feedback"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["business_id"], name: "index_businesses_hiring_invoice_requests_on_business_id"
   end
 
   create_table "conversations", force: :cascade do |t|
@@ -130,6 +145,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_03_19_212906) do
     t.boolean "source_contributor", default: false, null: false
     t.integer "response_rate", default: 0, null: false
     t.string "mastodon"
+    t.boolean "product_announcement_notifications", default: true
     t.index ["public_profile_key"], name: "index_developers_on_public_profile_key", unique: true
     t.index ["textsearchable_index_col"], name: "textsearchable_index", using: :gin
     t.index ["user_id"], name: "index_developers_on_user_id"
@@ -447,6 +463,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_03_19_212906) do
 
   add_foreign_key "active_storage_attachments", "active_storage_blobs", column: "blob_id"
   add_foreign_key "active_storage_variant_records", "active_storage_blobs", column: "blob_id"
+  add_foreign_key "businesses_hiring_invoice_requests", "businesses"
   add_foreign_key "hired_forms", "developers"
   add_foreign_key "hiring_agreements_signatures", "hiring_agreements_terms"
   add_foreign_key "hiring_agreements_signatures", "users"