+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### March
 
+* March 23 - Business survey email + opt-out checkbox #830
+* March 21 - Added business new hire forms and admin new hire forms #818 @RomanTurner @MikeRayUX
+* March 21 - Announce specialties to developers #817 @sarvaiyanidhi
 * March 20 - Update developer query to not duplicate records with multiple matching specialties #822 @bschrag620
 * March 18 - Add developer specialties #803
 * March 16 - Don't retry emails for invalid/unsubscribed recipients #798 @zzJZzz